fixed 5/9 with double

+++ Unit-02/src/Fahrenheit.java
@@ -7,7 +7,7 @@ public class Fahrenheit
 {
 	public static double convert( double f )
 	{
-		double cel = (5/9)*(f-32);
+		double cel = ((double)5/9)*(f-32);
     return cel;
 	}
 }
